refactor(persistence): move findMostRecentRetrospectiveFor into the memory plugin
The fork-chain walk that finds a conversation's most recent memory-retrospective
is memory-domain logic — only the memory-retrospective job + tests call it, and
it was the sole reason conversation-crud imported MEMORY_RETROSPECTIVE_SOURCES.
Move it (verbatim, with MAX_FORK_CHAIN_DEPTH) into the memory plugin. The query
reads the conversations table via persistence helpers (memory -> persistence).
conversation-crud now imports zero memory modules; its layering-allowlist entry
is removed entirely. Behavior byte-identical.
