+++ assistant/src/__tests__/background-workers-disk-pressure.test.ts
@@ -99,6 +99,12 @@ mock.module("../daemon/process-message.js", () => ({
 }));
 
 const createdConversations: Array<{ conversationType: string }> = [];
+mock.module(
+  "../plugins/defaults/memory/find-most-recent-retrospective-for.js",
+  () => ({
+    findMostRecentRetrospectiveFor: mock(() => null),
+  }),
+);
 mock.module("../persistence/conversation-crud.js", () => ({
   setConversationProcessingStartedAt: () => {},
   isConversationProcessing: () => false,
@@ -122,7 +128,6 @@ mock.module("../persistence/conversation-crud.js", () => ({
   })),
   deleteLastExchange: mock(() => 0),
   findAnalysisConversationFor: mock(() => null),
-  findMostRecentRetrospectiveFor: mock(() => null),
   forkConversation: mock(() => ({ id: "conv-fork" })),
   forkConversationForRetrospective: mock(async () => ({ id: "conv-fork" })),
   getConversationOverrideProfile: () => undefined,
