+++ assistant/src/plugins/defaults/memory/memory-retrospective-job.ts
@@ -57,7 +57,6 @@ import {
   type ConversationRow,
   deleteConversation,
   deleteConversationGently,
-  findMostRecentRetrospectiveFor,
   forkConversationForRetrospective,
   getConversation,
   getMessagesAfter,
@@ -73,6 +72,7 @@ import { resolveUserSlug } from "../../../prompts/persona-resolver.js";
 import type { SystemPromptPersonaOverride } from "../../../prompts/system-prompt.js";
 import { wakeAgentForOpportunity } from "../../../runtime/agent-wake.js";
 import { getLogger } from "../../../util/logger.js";
+import { findMostRecentRetrospectiveFor } from "./find-most-recent-retrospective-for.js";
 import {
   MEMORY_RETROSPECTIVE_FORK_SOURCE,
   MEMORY_RETROSPECTIVE_GROUP_ID,
